Under-the-hood changes and design tweaks
I've changed the way that the osc function works so it can be used for
LFOs too. I'm still not happy with the way that the functions play and
flow work, but I'll redo them once I add a sequencer.

Some of the CSS was a bit dodgy - non-centred text in buttons and the
dropdowns not appearing correctly on Windows. This has now been fixed.

Next update will include ADSR and a volume slider. Possibly also an LFO.